+++ mini_projects/day-06-random.py
@@ -1,0 +1,5 @@
+import random
+
+print(random.choice(["Heads","Tails"]))
+
+print(random.randint(1,6))